+++ packages/react-dom/src/server/ReactDOMServerFormatConfig.js
