+++ packages/react-native-renderer/src/server/ReactNativeServerFormatConfig.js
